+++ extension-impl/codec-sofa-hessian/src/main/java/com/alipay/sofa/rpc/codec/sofahessian/serialize/AbstractRpcSerializeObjector.java
@@ -14,12 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.codec.sofahessian.decoder;
+package com.alipay.sofa.rpc.codec.sofahessian.serialize;
 
 import com.alipay.hessian.ClassNameResolver;
 import com.alipay.hessian.NameBlackListFilter;
-import com.alipay.hessian.generic.model.GenericObject;
-import com.alipay.sofa.rpc.codec.RpcDecoderObjector;
 import com.alipay.sofa.rpc.codec.sofahessian.BlackListFileLoader;
 import com.alipay.sofa.rpc.codec.sofahessian.GenericMultipleClassLoaderSofaSerializerFactory;
 import com.alipay.sofa.rpc.codec.sofahessian.GenericSingleClassLoaderSofaSerializerFactory;
@@ -30,26 +28,16 @@ import com.alipay.sofa.rpc.common.RpcConfigs;
 import com.alipay.sofa.rpc.common.RpcOptions;
 import com.alipay.sofa.rpc.common.SofaConfigs;
 import com.alipay.sofa.rpc.common.SofaOptions;
-import com.alipay.sofa.rpc.common.struct.UnsafeByteArrayInputStream;
-import com.alipay.sofa.rpc.common.utils.ClassTypeUtils;
-import com.alipay.sofa.rpc.config.ConfigUniqueNameGenerator;
 import com.alipay.sofa.rpc.context.RpcInternalContext;
 import com.alipay.sofa.rpc.core.exception.RpcErrorType;
 import com.alipay.sofa.rpc.core.exception.SofaRpcException;
-import com.alipay.sofa.rpc.core.request.SofaRequest;
-import com.alipay.sofa.rpc.core.response.SofaResponse;
-import com.alipay.sofa.rpc.transport.AbstractByteBuf;
-import com.caucho.hessian.io.Hessian2Input;
 import com.caucho.hessian.io.SerializerFactory;
-
-import java.io.IOException;
-import java.util.Map;
 
 /**
  * @author bystander
- * @version $Id: SofaRequestRpcDecoderObjector.java, v 0.1 2018年09月07日 4:34 PM bystander Exp $
+ * @version $Id: AbstractRpcSerializeObjector.java, v 0.1 2018年09月08日 7:15 PM bystander Exp $
  */
-public class SofaResponseRpcDecoderObjector implements RpcDecoderObjector<SofaResponse> {
+public class AbstractRpcSerializeObjector {
 
     /**
      * Normal Serializer Factory
@@ -60,10 +48,8 @@ public class SofaResponseRpcDecoderObjector implements RpcDecoderObjector<SofaRe
      */
     protected SerializerFactory genericSerializerFactory;
 
-    /**
-     * Instantiates a new Sofa hessian serializer.
-     */
-    public SofaResponseRpcDecoderObjector() {
+    public AbstractRpcSerializeObjector() {
+
         boolean enableMultipleClassLoader = RpcConfigs.getBooleanValue(RpcOptions.MULTIPLE_CLASSLOADER_ENABLE);
         serializerFactory = getSerializerFactory(enableMultipleClassLoader, false);
         genericSerializerFactory = getSerializerFactory(enableMultipleClassLoader, true);
@@ -74,73 +60,6 @@ public class SofaResponseRpcDecoderObjector implements RpcDecoderObjector<SofaRe
             serializerFactory.setClassNameResolver(resolver);
             genericSerializerFactory.setClassNameResolver(resolver);
         }
-    }
-
-    @Override
-    public void decodeObjectByTemplate(AbstractByteBuf data, Map<String, String> context, SofaResponse template)
-        throws SofaRpcException {
-        try {
-            UnsafeByteArrayInputStream inputStream = new UnsafeByteArrayInputStream(data.array());
-            Hessian2Input input = new Hessian2Input(inputStream);
-            // 根据SerializeType信息决定序列化器
-            boolean genericSerialize = context != null && isGenericResponse(
-                context.get(RemotingConstants.HEAD_GENERIC_TYPE));
-            if (genericSerialize) {
-                input.setSerializerFactory(genericSerializerFactory);
-                GenericObject genericObject = (GenericObject) input.readObject();
-                template.setErrorMsg((String) genericObject.getField("errorMsg"));
-                template.setAppResponse(genericObject.getField("appResponse"));
-                template.setResponseProps((Map<String, String>) genericObject.getField("responseProps"));
-            } else {
-                input.setSerializerFactory(serializerFactory);
-                SofaResponse tmp = (SofaResponse) input.readObject();
-                // copy values to template
-                template.setErrorMsg(tmp.getErrorMsg());
-                template.setAppResponse(tmp.getAppResponse());
-                template.setResponseProps(tmp.getResponseProps());
-            }
-            input.close();
-        } catch (IOException e) {
-            throw buildDeserializeError(e.getMessage(), e);
-        }
-    }
-
-    @Override
-    public SofaResponse decodeObject(AbstractByteBuf data, Map<String, String> context) throws SofaRpcException {
-        try {
-            UnsafeByteArrayInputStream inputStream = new UnsafeByteArrayInputStream(data.array());
-            Hessian2Input input = new Hessian2Input(inputStream);
-            // 根据SerializeType信息决定序列化器
-            Object object;
-            boolean genericSerialize = context != null && isGenericResponse(
-                context.get(RemotingConstants.HEAD_GENERIC_TYPE));
-            if (genericSerialize) {
-                input.setSerializerFactory(genericSerializerFactory);
-                GenericObject genericObject = (GenericObject) input.readObject();
-                SofaResponse sofaResponse = new SofaResponse();
-                sofaResponse.setErrorMsg((String) genericObject.getField("errorMsg"));
-                sofaResponse.setAppResponse(genericObject.getField("appResponse"));
-                sofaResponse.setResponseProps((Map<String, String>) genericObject.getField("responseProps"));
-                object = sofaResponse;
-            } else {
-                input.setSerializerFactory(serializerFactory);
-                object = input.readObject();
-            }
-            input.close();
-            return (SofaResponse) object;
-        } catch (IOException e) {
-            throw buildDeserializeError(e.getMessage(), e);
-        }
-    }
-
-    /**
-     * Is generic response boolean.
-     *
-     * @param serializeType the serialize type
-     * @return the boolean
-     */
-    protected boolean isGenericResponse(String serializeType) {
-        return serializeType != null && serializeType.equals(RemotingConstants.SERIALIZE_FACTORY_GENERIC);
     }
 
     /**
@@ -158,6 +77,10 @@ public class SofaResponseRpcDecoderObjector implements RpcDecoderObjector<SofaRe
             return multipleClassLoader ? new MultipleClassLoaderSofaSerializerFactory() :
                 new SingleClassLoaderSofaSerializerFactory();
         }
+    }
+
+    protected SofaRpcException buildDeserializeError(String message) {
+        return new SofaRpcException(getErrorCode(false), message);
     }
 
     protected SofaRpcException buildSerializeError(String message, Throwable throwable) {
@@ -181,8 +104,23 @@ public class SofaResponseRpcDecoderObjector implements RpcDecoderObjector<SofaRe
         }
     }
 
-    protected SofaRpcException buildDeserializeError(String message) {
-        return new SofaRpcException(getErrorCode(false), message);
+    /**
+     * Is generic request boolean.
+     *
+     * @param serializeType the serialize type
+     * @return the boolean
+     */
+    protected boolean isGenericRequest(String serializeType) {
+        return serializeType != null && !serializeType.equals(RemotingConstants.SERIALIZE_FACTORY_NORMAL);
     }
 
+    /**
+     * Is generic response boolean.
+     *
+     * @param serializeType the serialize type
+     * @return the boolean
+     */
+    protected boolean isGenericResponse(String serializeType) {
+        return serializeType != null && serializeType.equals(RemotingConstants.SERIALIZE_FACTORY_GENERIC);
+    }
 }
