+++ extension-impl/codec-sofa-hessian/src/main/java/com/alipay/sofa/rpc/codec/sofahessian/serialize/SofaRequestRpcSerializeObjector.java
@@ -14,64 +14,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.rpc.codec.sofahessian.decoder;
+package com.alipay.sofa.rpc.codec.sofahessian.serialize;
 
-import com.alipay.hessian.ClassNameResolver;
-import com.alipay.hessian.NameBlackListFilter;
-import com.alipay.sofa.rpc.codec.RpcDecoderObjector;
-import com.alipay.sofa.rpc.codec.sofahessian.BlackListFileLoader;
-import com.alipay.sofa.rpc.codec.sofahessian.GenericMultipleClassLoaderSofaSerializerFactory;
-import com.alipay.sofa.rpc.codec.sofahessian.GenericSingleClassLoaderSofaSerializerFactory;
-import com.alipay.sofa.rpc.codec.sofahessian.MultipleClassLoaderSofaSerializerFactory;
-import com.alipay.sofa.rpc.codec.sofahessian.SingleClassLoaderSofaSerializerFactory;
+import com.alipay.sofa.rpc.codec.RpcSerializeObjector;
 import com.alipay.sofa.rpc.common.RemotingConstants;
-import com.alipay.sofa.rpc.common.RpcConfigs;
-import com.alipay.sofa.rpc.common.RpcOptions;
-import com.alipay.sofa.rpc.common.SofaConfigs;
-import com.alipay.sofa.rpc.common.SofaOptions;
 import com.alipay.sofa.rpc.common.struct.UnsafeByteArrayInputStream;
+import com.alipay.sofa.rpc.common.struct.UnsafeByteArrayOutputStream;
 import com.alipay.sofa.rpc.common.utils.ClassTypeUtils;
 import com.alipay.sofa.rpc.config.ConfigUniqueNameGenerator;
-import com.alipay.sofa.rpc.context.RpcInternalContext;
-import com.alipay.sofa.rpc.core.exception.RpcErrorType;
 import com.alipay.sofa.rpc.core.exception.SofaRpcException;
 import com.alipay.sofa.rpc.core.request.SofaRequest;
 import com.alipay.sofa.rpc.transport.AbstractByteBuf;
+import com.alipay.sofa.rpc.transport.ByteStreamWrapperByteBuf;
 import com.caucho.hessian.io.Hessian2Input;
-import com.caucho.hessian.io.SerializerFactory;
+import com.caucho.hessian.io.Hessian2Output;
 
 import java.io.IOException;
 import java.util.Map;
 
 /**
  * @author bystander
- * @version $Id: SofaRequestRpcDecoderObjector.java, v 0.1 2018年09月07日 4:34 PM bystander Exp $
+ * @version $Id: SofaRequestRpcSerializeObjector.java, v 0.1 2018年09月07日 4:34 PM bystander Exp $
  */
-public class SofaRequestRpcDecoderObjector implements RpcDecoderObjector<SofaRequest> {
-
-    /**
-     * Normal Serializer Factory
-     */
-    protected SerializerFactory serializerFactory;
-    /**
-     * Generic Serializer Factory
-     */
-    protected SerializerFactory genericSerializerFactory;
+public class SofaRequestRpcSerializeObjector extends AbstractRpcSerializeObjector implements
+                                                                                 RpcSerializeObjector<SofaRequest> {
 
     /**
      * Instantiates a new Sofa hessian serializer.
      */
-    public SofaRequestRpcDecoderObjector() {
-        boolean enableMultipleClassLoader = RpcConfigs.getBooleanValue(RpcOptions.MULTIPLE_CLASSLOADER_ENABLE);
-        serializerFactory = getSerializerFactory(enableMultipleClassLoader, false);
-        genericSerializerFactory = getSerializerFactory(enableMultipleClassLoader, true);
-        if (RpcConfigs.getBooleanValue(RpcOptions.SERIALIZE_BLACKLIST_ENABLE) &&
-            SofaConfigs.getBooleanValue(SofaOptions.CONFIG_SERIALIZE_BLACKLIST, true)) {
-            ClassNameResolver resolver = new ClassNameResolver();
-            resolver.addFilter(new NameBlackListFilter(BlackListFileLoader.SOFA_SERIALIZE_BLACK_LIST, 8192));
-            serializerFactory.setClassNameResolver(resolver);
-            genericSerializerFactory.setClassNameResolver(resolver);
-        }
+    public SofaRequestRpcSerializeObjector() {
+        super();
     }
 
     @Override
@@ -141,56 +113,33 @@ public class SofaRequestRpcDecoderObjector implements RpcDecoderObjector<SofaReq
         }
     }
 
-    /**
-     * Is generic response boolean.
-     *
-     * @param serializeType the serialize type
-     * @return the boolean
-     */
-    protected boolean isGenericResponse(String serializeType) {
-        return serializeType != null && serializeType.equals(RemotingConstants.SERIALIZE_FACTORY_GENERIC);
-    }
+    @Override
+    public AbstractByteBuf encodeObject(SofaRequest sofaRequest, Map<String, String> context) {
+        try {
+            UnsafeByteArrayOutputStream outputStream = new UnsafeByteArrayOutputStream();
+            Hessian2Output output = new Hessian2Output(outputStream);
 
-    /**
-     * Gets serializer factory.
-     *
-     * @param multipleClassLoader the multiple class loader
-     * @param generic             the generic
-     * @return the serializer factory
-     */
-    protected SerializerFactory getSerializerFactory(boolean multipleClassLoader, boolean generic) {
-        if (generic) {
-            return multipleClassLoader ? new GenericMultipleClassLoaderSofaSerializerFactory() :
-                new GenericSingleClassLoaderSofaSerializerFactory();
-        } else {
-            return multipleClassLoader ? new MultipleClassLoaderSofaSerializerFactory() :
-                new SingleClassLoaderSofaSerializerFactory();
+            // 根据SerializeType信息决定序列化器
+            boolean genericSerialize = context != null &&
+                isGenericRequest(context.get(RemotingConstants.HEAD_GENERIC_TYPE));
+            if (genericSerialize) {
+                output.setSerializerFactory(genericSerializerFactory);
+            } else {
+                output.setSerializerFactory(serializerFactory);
+            }
+
+            output.writeObject(sofaRequest);
+            final Object[] args = sofaRequest.getMethodArgs();
+            if (args != null) {
+                for (Object arg : args) {
+                    output.writeObject(arg);
+                }
+            }
+            output.close();
+
+            return new ByteStreamWrapperByteBuf(outputStream);
+        } catch (IOException e) {
+            throw buildSerializeError(e.getMessage(), e);
         }
     }
-
-    protected SofaRpcException buildSerializeError(String message, Throwable throwable) {
-        return new SofaRpcException(getErrorCode(true), message, throwable);
-    }
-
-    protected SofaRpcException buildDeserializeError(String message, Throwable throwable) {
-        return new SofaRpcException(getErrorCode(false), message, throwable);
-    }
-
-    /**
-     * @param serialize true is serialize, false is deserialize.
-     */
-    private int getErrorCode(boolean serialize) {
-        if (RpcInternalContext.getContext().isProviderSide()) {
-            return serialize ? RpcErrorType.SERVER_SERIALIZE : RpcErrorType.SERVER_DESERIALIZE;
-        } else if (RpcInternalContext.getContext().isConsumerSide()) {
-            return serialize ? RpcErrorType.CLIENT_SERIALIZE : RpcErrorType.CLIENT_DESERIALIZE;
-        } else {
-            return RpcErrorType.UNKNOWN;
-        }
-    }
-
-    protected SofaRpcException buildDeserializeError(String message) {
-        return new SofaRpcException(getErrorCode(false), message);
-    }
-
 }
